In addition to local "hepsiburada"  shopping site code, "Amazon" global shopping site code is added to the repo.
For each shopping site both serial and parallel version of the webscraper codes are published so that; for people who do not aim to run the codes parallel can use the serial versions.
Parallel versions use MPI and have load balancing mechanisms to distribute the load (work) to the processes equally.

README files are updated with more comprehensive explanations and output files are added for the example runs of the scripts (.csv files).